move encoding value/numbers into IndexLevel
